LoggingSystem/requirements.txt
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
pycparser==2.20
pycrypto==2.6.1
pyOpenSSL==19.1.0
six==1.15.0
Werkzeug==1.0.1
